service Test and Repository Test its finished

+++ src/main/java/com/example/filmDbProject/Service/ActorServiceImpl.java
@@ -5,35 +5,64 @@ import com.example.filmDbProject.Repository.ActorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+import java.util.Optional;
+
 @Service
 public class ActorServiceImpl implements ActorService{
     @Autowired
     ActorRepository actorRepository;
 
     @Override
-    public Actor getActor(int idActor) {
-        Actor actor = actorRepository.findById(idActor).get();
-        return actor;
+    public Actor getActor(Integer idActor) {
+        if (idActor == null){
+            return null;
+        }
+        Optional<Actor> actor = actorRepository.findById(idActor);
+        if (actor.isPresent()){
+            return actor.get();
+        }
+        return null;
     }
 
     @Override
     public Actor saveActor(Actor actor) {
+        if (actor == null){
+            return null;
+        }
         return actorRepository.save(actor);
     }
 
     @Override
-    public void delete(int id) {
-        actorRepository.deleteById(id);
+    public void delete(Integer id) {
+        if (id == null){
+
+        }else {
+            actorRepository.deleteById(id);
+        }
+
     }
 
     @Override
-    public Actor updateActor(int idActor, Actor actor) {
-        Actor actor1 = actorRepository.findById(idActor).get();
-        actor1.setLastUpdate(actor.getLastUpdate());
-        actor1.setLastName(actor.getLastName());
-        actor1.setFirstName(actor.getFirstName());
+    public Actor updateActor(Integer idActor, Actor actor) {
+        if (idActor==null || actor == null){
+            return null;
+        }
 
+        Optional<Actor> actor1 = actorRepository.findById(idActor);
+        if (actor1.isPresent()){
+            Actor actorGet = actor1.get();
+            actorGet.setLastUpdate(actor.getLastUpdate());
+            actorGet.setLastName(actor.getLastName());
+            actorGet.setFirstName(actor.getFirstName());
+            return actorRepository.save(actorGet);
+        }else {
+            return null;
+        }
+    }
 
-        return actorRepository.save(actor1);
+    @Override
+    public List<Actor> getActors() {
+        return (List<Actor>) actorRepository.findAll();
     }
 }
